Ordner feilen med ./suuid.t --todo der $Outdir og $Outfile ikke var definert og lagde feilmelding i suuid/tests/log/toto-tests.log .
* /trunk/src/suuid/tests/suuid.t
  Flytter defineringa av $Outdir før den hopper til todo_section.
  Når den rensker opp i filene på slutten, sjekker den om $Outfile er 
  definert før den prøver å slette.

* /trunk/src/suuid/tests/log/todo-tests.log
  Feilmeldingene er borte. Ingen tester er kjørt.

51513b68-fb7a-11dd-8300-000475e441b9


git-svn-id: svn+ssh://svnhost/utils@2938 ddca7c6a-6bd8-0310-abd8-a0f6a7c394c1

+++ src/suuid/tests/suuid.t
@@ -72,6 +72,14 @@ diag(sprintf("========== Executing \"%s%s%s\" ==========",
     scalar(@cmdline_array) ? " " : "",
     join(" ", @cmdline_array)));
 
+my $Outdir = "tmp-suuid-t-$$-" . substr(rand, 2, 8);
+if (-e $Outdir) {
+    die("$progname: $Outdir: WTF?? Directory element already exists.");
+}
+unless (mkdir($Outdir)) {
+    die("$progname: $Outdir: Cannot mkdir(): $!\n");
+}
+
 if ($Opt{'todo'} && !$Opt{'all'}) {
     goto todo_section;
 }
@@ -89,14 +97,6 @@ END
 # }}}
 
 =cut
-
-my $Outdir = "tmp-suuid-t-$$-" . substr(rand, 2, 8);
-if (-e $Outdir) {
-    die("$progname: $Outdir: WTF?? Directory element already exists.");
-}
-unless (mkdir($Outdir)) {
-    die("$progname: $Outdir: Cannot mkdir(): $!\n");
-}
 
 diag("Testing -h (--help) option...");
 likecmd("$CMD -h", # {{{
@@ -338,7 +338,9 @@ local $TODO = "";
 
 diag("Testing finished.");
 
-unlink($Outfile) || warn("$progname: $Outfile: Cannot delete file: $!\n");
+if (defined($Outfile)) {
+    unlink($Outfile) || warn("$progname: $Outfile: Cannot delete file: $!\n");
+}
 rmdir($Outdir) || warn("$progname: $Outdir: Cannot remove directory: $!\n");
 
 sub testcmd {
